Moved to lib for bytecode compilation.
